+++ R/jam_shrink_df.R
@@ -9,6 +9,66 @@
 #'
 #' @family jamses utilities
 #'
+#' @param x `data.frame` (or equivalent)
+#' @param groupBy `character` vector with one of the following:
+#'    * one or more columns in `colnames(x)`. The values in these columns
+#'    will define the row groups used.
+#'    * `character` or `factor` with length equal to `nrow(x)`. These
+#'    values will define the row groups used.
+#' @param stringShrinkFunc `function`, default uses `jamba::cPasteSU()`,
+#'    used for `character` or `factor` columns. Note that string
+#'    columns are handled differently than `numeric` columns by applying
+#'    vectorized operations across the complete set of rows in one step,
+#'    rather than calling `data.table` on each subgroup.
+#' @param numShrinkFunc `function`, default `function(x)mean(x, na.rm=TRUE)`,
+#'    used for `numeric` columns. Note that this function is applied
+#'    to each row group by `data.table`, and is typically very efficient
+#'    for `numeric` values.
+#' @param addStringCols `character` with optional `numeric` columns that
+#'    should be handled as if they were `character` columns. Default `NULL`.
+#' @param stringToNumFunc `function` used for `addStringCols` when converting
+#'    `numeric` columns to `character`. Default `as.character()` retains
+#'    the full `numeric` value, however it may be useful to use something
+#'    like `function(x)signif(x, digits=3)` to limit the output to
+#'    only three significant digits, or `function(x)format(x, digits=3)`.
+#' @param keepNAgroupBy `logical`, default TRUE, whether to convert `NA`
+#'    values in row groups to `""` so they are retained in the output.
+#' @param includeNumReps `logical` indicating whether to add a column
+#'    `"numReps"` to the output, with the `integer` number of rows
+#'    in each row group.
+#' @param collapseMethod `integer` default 2, indicating the internal
+#'    collapse method used. Experimental.
+#'    * `1` collapses each `numeric` column independently.
+#'    * `2` collapses each set of `numeric` columns that use the same
+#'    numeric shrink function. When all `numeric` columns use the same
+#'    shrink function, they are all calculated in a single step, which
+#'    is typically much faster.
+#' @param applyKeepCols `logical` whether to reorder columns to match the
+#'    input `colnames(x)`.
+#' @param verbose `logical` indicating whether to print verbose output.
+#' @param ... additional arguments are ignored.
+#'
+#' @examples
+#' testdf <- data.frame(check.names=FALSE,
+#'    SYMBOL=rep(c("ACTB", "GAPDH", "PPIA"), c(2, 3, 1)),
+#'    `logFC B-A`=c(1.4, 1.4, 2.3, NA, 2.3, 5.1),
+#'    probe=paste0("probe", 1:6))
+#' shrink_df(testdf, by="SYMBOL", num_func=function(x){mean(x, na.rm=TRUE)})
+#'
+#' testdftall <- do.call(rbind, lapply(1:10000, function(i){
+#'    idf <- testdf;
+#'    idf$SYMBOL <- paste0(idf$SYMBOL, "_", i);
+#'    idf;
+#' }))
+#' head(testdftall, 12)
+#' shrunk_tall <- shrink_df(testdftall,
+#'    by="SYMBOL",
+#'    num_func=function(x){mean(x, na.rm=TRUE)})
+#'
+#' shrunk_tall2 <- jamses::shrinkDataFrame(testdftall,
+#'    groupBy="SYMBOL")
+#' head(shrunk_tall2, 12)
+#'
 #' @export
 shrinkDataFrame <- function
 (x,
@@ -82,7 +142,7 @@ shrinkDataFrame <- function
    if (length(groupBy) == nrow(iTopTableEx)) {
       ## groupBy use as-is
       ## If factor, make sure NA is a level if there are NAs
-      if (igrepHas("factor", class(groupBy))) {
+      if (jamba::igrepHas("factor", class(groupBy))) {
          groupBy <- addNA(groupBy, ifany=TRUE);
       }
       if (!keepNAgroupBy && any(is.na(as.character(groupBy)))) {
@@ -100,7 +160,8 @@ shrinkDataFrame <- function
    } else {
       if (any(groupBy %in% colnames(iTopTableEx))) {
          groupByCols <- intersect(groupBy, colnames(iTopTableEx));
-         groupBy <- pasteByRow(iTopTableEx[,groupByCols,drop=FALSE], sep="::");
+         groupBy <- jamba::pasteByRow(
+            iTopTableEx[, groupByCols, drop=FALSE], sep="::");
          if (verbose) {
             jamba::printDebug("shrinkDataFrame(): ",
                "head(groupByCols):",
@@ -109,10 +170,14 @@ shrinkDataFrame <- function
                "head(groupBy):",
                head(groupBy));
          }
-         groupByDF <- data.frame(check.names=FALSE, stringsAsFactors=FALSE,
-            iTopTableEx[match(unique(groupBy), groupBy),groupByCols,drop=FALSE], row.names=unique(groupBy));
+         matchg <- match(unique(groupBy), groupBy);
+         groupByDF <- data.frame(check.names=FALSE,
+            stringsAsFactors=FALSE,
+            iTopTableEx[matchg, groupByCols, drop=FALSE],
+            row.names=unique(groupBy));
       } else {
-         stop("groupBy must either be a vector equal to nrow(x), or a vector of colnames contained in x.");
+         stop(paste0("groupBy must either be a vector equal to nrow(x),",
+            " or a vector of colnames contained in x."));
       }
    }
    if (any(is.na(as.character(groupBy)))) {
@@ -137,11 +202,11 @@ shrinkDataFrame <- function
 
    ## Detect numeric columns
    numCols <- names(colClasses)[sapply(colClasses, function(iClass){
-      igrepHas("integer|numeric|float", iClass);
+      jamba::igrepHas("integer|numeric|float", iClass);
    })];
    ## Detect string/character
    stringCols <- names(colClasses)[sapply(colClasses, function(iClass){
-      igrepHas("character|factor|ordered|logical", iClass);
+      jamba::igrepHas("character|factor|ordered|logical", iClass);
    })];
    ## All other classes are ignored
    if (verbose) {
@@ -182,7 +247,8 @@ shrinkDataFrame <- function
             ## we have numCols.
             iNewCols <- setdiff(numCols, names(numShrinkFunc));
             numShrinkFunc <- c(numShrinkFunc,
-               jamba::nameVector(rep(numShrinkFunc, length.out=length(iNewCols)),
+               jamba::nameVector(
+                  rep(numShrinkFunc, length.out=length(iNewCols)),
                   iNewCols));
          }
          numShrinkFunc <- numShrinkFunc[numCols];
@@ -273,18 +339,18 @@ shrinkDataFrame <- function
                   "iCol:",
                   iCol,
                   fgText=c("orange", "lightgreen"));
+               # jamba::printDebug("   shrinkDataFrame(): ",
+               #    "colnames(iTopTableEx): ",
+               #    colnames(iTopTableEx));
                jamba::printDebug("   shrinkDataFrame(): ",
-                  "colnames(iTopTableEx): ",
-                  colnames(iTopTableEx));
+                  "head(data.frame):");
                print(head(
                   data.frame(iTopTableEx[, iCol, drop=FALSE],
                      check.names=FALSE,
                      stringsAsFactors=FALSE)
                ));
-               jamba::printDebug("   shrinkDataFrame(): ",
-                  "printed head(data.frame) above.");
             }
-            grOLi <- splicejam::shrinkMatrix(
+            grOLi <- shrink_matrix(
                data.frame(iTopTableEx[, iCol, drop=FALSE],
                   check.names=FALSE,
                   stringsAsFactors=FALSE),
@@ -302,24 +368,29 @@ shrinkDataFrame <- function
                   "collapseMethod:", collapseMethod,
                   " running only once for all numeric values.");
             }
-            numShrunk <- splicejam::shrinkMatrix(as.data.frame(iTopTableEx[,numCols,drop=FALSE]),
+            numShrunk <- shrink_matrix(
+               as.data.frame(iTopTableEx[, numCols, drop=FALSE]),
                groupBy=rmNA(naValue="NA", groupBy),
-               shrinkFunc=numShrinkFunc[[1]], returnClass="matrix");
+               shrinkFunc=numShrinkFunc[[1]],
+               returnClass="matrix");
          } else {
             if (verbose) {
                jamba::printDebug("shrinkDataFrame(): ",
                   "collapseMethod:", collapseMethod,
                   " running on each numeric column.");
             }
-            numShrunk <- do.call(cbind, lapply(jamba::nameVector(numCols), function(iCol){
-               if (verbose) {
-                  jamba::printDebug("  - ", iCol, c("orange", "lightgreen"));
-                  print(numShrinkFunc[[iCol]]);
-               }
-               grOLi <- splicejam::shrinkMatrix(as.data.frame(iTopTableEx[,iCol,drop=FALSE]),
-                  groupBy=groupBy,
-                  shrinkFunc=numShrinkFunc[[iCol]], returnClass="matrix");
-            }));
+            numShrunk <- do.call(cbind,
+               lapply(jamba::nameVector(numCols), function(iCol){
+                  if (verbose) {
+                     jamba::printDebug("  - ", iCol,
+                        fgText=c("orange", "lightgreen"));
+                     print(numShrinkFunc[[iCol]]);
+                  }
+                  grOLi <- shrink_matrix(
+                     as.data.frame(iTopTableEx[, iCol, drop=FALSE]),
+                     groupBy=groupBy,
+                     shrinkFunc=numShrinkFunc[[iCol]], returnClass="matrix");
+               }));
          }
       }
       numShrunkDF <- data.frame(check.names=FALSE,
@@ -333,7 +404,7 @@ shrinkDataFrame <- function
       iMatch <- match(unique(groupBy), rownames(numShrunkDF));
       icolnames1 <- colnames(groupByDF);
       groupByDF <- cbind(groupByDF,
-         numShrunkDF[iMatch,,drop=FALSE]);
+         numShrunkDF[iMatch, , drop=FALSE]);
       colnames(groupByDF) <- c(icolnames1,
          colnames(numShrunkDF));
    }
@@ -347,44 +418,48 @@ shrinkDataFrame <- function
             stringCols);
       }
       options("stringsAsFactor"=FALSE);
-      stringShrunkDF <- as.data.frame(lapply(jamba::nameVector(stringCols), function(iCol){
-         if (verbose) {
-            jamba::printDebug("   shrinkDataFrame(): ",
-               "stringShrinkFunc[['", iCol, "']]",
-               fgText=c("orange", "lightgreen"));
-         }
-         ## split vector into a list, then run the function on the list
-         ## which is intended to generate a vector of values
-         f1 <- iTopTableEx[[iCol]];
-         iValsL <- stringShrinkFunc[[iCol]](split(as.character(f1), groupBy));
-         if (igrepHas("factor", class(iTopTableEx[[iCol]]))) {
+      stringShrunkDF <- as.data.frame(
+         lapply(jamba::nameVector(stringCols), function(iCol){
             if (verbose) {
-               jamba::printDebug("      shrinkDataFrame(): ",
-                  "shrinking factor column values.");
+               jamba::printDebug("   shrinkDataFrame(): ",
+                  "stringShrinkFunc[['", iCol, "']]",
+                  fgText=c("orange", "lightgreen"));
             }
-            ## Generate new factor levels in the original order, followed by
-            ## sorted values, then keeping only those levels contained in
-            ## the resulting data
-            iLevels <- intersect(unique(c(levels(f1), mixedSort(iValsL))),
-               iValsL);
-            if (!NA %in% levels(iTopTableEx[[iCol]])) {
-               ## Bonus points: if NA was not in the levels previously, keep it out
-               ## otherwise NA will be a formal factor level
-               iLevels <- rmNA(iLevels);
+            ## split vector into a list, then run the function on the list
+            ## which is intended to generate a vector of values
+            f1 <- iTopTableEx[[iCol]];
+            iValsL <- stringShrinkFunc[[iCol]](split(as.character(f1),
+               groupBy));
+            if (jamba::igrepHas("factor", class(iTopTableEx[[iCol]]))) {
+               if (verbose) {
+                  jamba::printDebug("      shrinkDataFrame(): ",
+                     "shrinking factor column values.");
+               }
+               ## Generate new factor levels in the original order, followed by
+               ## sorted values, then keeping only those levels contained in
+               ## the resulting data
+               iLevels <- intersect(unique(c(levels(f1), mixedSort(iValsL))),
+                  iValsL);
+               if (!NA %in% levels(iTopTableEx[[iCol]])) {
+                  ## Bonus points: if NA was not in the levels previously, keep it out
+                  ## otherwise NA will be a formal factor level
+                  iLevels <- rmNA(iLevels);
+               }
+               iValsL <- factor(iValsL, levels=iLevels);
             }
-            iValsL <- factor(iValsL, levels=iLevels);
-         }
-         iValsL;
-      }));
+            iValsL;
+         }));
       colnames(stringShrunkDF) <- stringCols;
-      iMatch <- match(unique(groupBy), rownames(stringShrunkDF));
+      iMatch <- match(unique(groupBy),
+         rownames(stringShrunkDF));
       groupByDF <- cbind(groupByDF,
-         stringShrunkDF[iMatch,,drop=FALSE]);
+         stringShrunkDF[iMatch, , drop=FALSE]);
    }
    ## keepCols is used to protect against columns whose types are
    ## not considered num or string, and may be lost above
    if (applyKeepCols) {
-      keepCols <- intersect(colnames(iTopTableEx), colnames(groupByDF));
+      keepCols <- intersect(colnames(iTopTableEx),
+         colnames(groupByDF));
    } else {
       keepCols <- colnames(groupByDF);
    }